Add user information page for client and admin, add order history page for client

+++ CAS/CAS/Model/Dao/UserDao.cs
@@ -28,7 +28,7 @@ namespace Model.Dao
             return db.Users.SingleOrDefault(x => x.UserName == username);
         }
 
-        public User GetByID(int id)
+        public User GetByID(long id)
         {
             return db.Users.Find(id);
         }
